Creation de tous les controllers

+++ src/main/java/fr/utbm/fisa/communicationapi/infrastructure/repositories/DiscussionRepository.java
@@ -3,5 +3,5 @@ package fr.utbm.fisa.communicationapi.infrastructure.repositories;
 import fr.utbm.fisa.communicationapi.infrastructure.entities.Discussion;
 import org.springframework.data.repository.CrudRepository;
 
-public interface DiscussionRepository extends CrudRepository<Discussion, Long> {
+public interface DiscussionRepository extends CrudRepository<Discussion, Integer> {
 }
