+++ src/main/java/fr/utbm/fisa/communicationapi/infrastructure/repositories/MessageRepository.java
@@ -3,5 +3,5 @@ package fr.utbm.fisa.communicationapi.infrastructure.repositories;
 import fr.utbm.fisa.communicationapi.infrastructure.entities.Message;
 import org.springframework.data.repository.CrudRepository;
 
-public interface MessageRepository extends CrudRepository<Message, Long> {
+public interface MessageRepository extends CrudRepository<Message, Integer> {
 }
