+++ src/main/java/fr/utbm/fisa/communicationapi/infrastructure/repositories/PollRepository.java
@@ -3,5 +3,5 @@ package fr.utbm.fisa.communicationapi.infrastructure.repositories;
 import fr.utbm.fisa.communicationapi.infrastructure.entities.Poll;
 import org.springframework.data.repository.CrudRepository;
 
-public interface PollRepository extends CrudRepository<Poll, Long> {
+public interface PollRepository extends CrudRepository<Poll, Integer> {
 }
